Check R packages in dev container
As the comments explain, the R packaging situation is slightly
complicated.

We have additional packages due to the Rocker image. Some of these are
duplicates of those in the OpenSAFELY R image; some with the same
version and some with a different version.

An alternative approach might have been to:

* copy the OpenSAFELY R packages to a path that is unique
* add that library tree to the front of `.libPaths`,
  so these packages always take precedence over other installed versions
* when comparing, compare the `installed_packages` in
  the specific OpenSAFELY library tree of interest:

  ```r
  installed.packages(lib.loc="/usr/local/lib/R/opensafely-site-library")
  ```

  taken from the OpenSAFELY image. That would remove the need for the
  exemptions when checking.

Fow now, it's possibly more useful still being aware of these additional
packages and checking for them explicitly.