Consistently double quote shell variables
The lack of double quotes for these variables results in a `shellcheck`
warning:

https://www.shellcheck.net/wiki/SC2086

Since double quotes are now used elsewhere, it seems reasonable to add
them throughout for consistency in the edited file.

Quotes are not absolutely necessary given the specific values that are
being used here; there are no spaces or filename expansions there.
However, it's more difficult to reason that through for each variable,
than simply making them safe by quoting.