Add MOTD to be echoed in every bash shell
Many of our users may be unfamiliar with VS Code and
specifically using its integrated terminal.
These messages aim to reassure the users that this terminal
(which is open by default in Codespaces) that it behaves like
the terminal they would usually use for OpenSAFELY CLI
and to remind them where they can get help.