Add scheduled check of research-template container
This is a minimal workflow that tests the `research-template` dev
container daily, using whichever Docker image is specified in the
`devcontainer.json` in the `main` branch there.

It could be extended to run on every build
with the image to be tested.

This workflow should also report to Slack on failure, instead of whoever
last edited the cron line.