Add a workflow to test development Docker images
The use case for this manually run workflow is:

* You have made some changes to the Docker image build.
* You want to verify that the current dev container configuration works
  with that Docker image build, without first having to merge the image,
  and then run the existing dev container workflow against the published
  image, and only then find out that the dev container fails to function
  for some reason.

There is no easy way otherwise of testing an in-development Docker image
build with the dev container configuration. While the existing dev
container test workflow is useful, it only runs with the published image
(currently set to `v0`).

We originally had something like this when developing the tests, but
this was removed in favour of a schedule checking that the dev
container works with the published image. It may be that we do not want
this type of workflow running on every commit/PR push, but is useful as
extra verification before merging, when extensive changes are made.