Fix changes in r:v1 image packaging
We switched the r:v1 package metadata to use a markdown file rather than
a csv, which broke this test, as we removed the csv in the r-docker
repo.

This change updates the test to use the renv.lock file instead, and with
a bit of trickery, generates an renv.lock file from the devcontainer to
compare it against.

Also improved the test output when failing to tell the developer
specifically that extra R packages were found, and which ones.