Use opensafely launch rstudio to provide rstudio
This includes detecting which version of the r image a project is
using, and launching the correct version.

The R autocompletion in vscode editor is still based on the dev
container version of R, which is the v1.

This frees us from needing to use Rocker as a base image to get
rstudio, but changing that is a later task.

Fixes #89